+++ app/src/main/java/com/playgilround/schedule/client/realm/ScheduleR.java
@@ -41,8 +41,8 @@ public class ScheduleR extends RealmObject {
     private int day;
     private int eventSetId;
 
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
     private String hTime;
 
     private String sendServer; //서버 전송여부
@@ -203,19 +203,19 @@ public class ScheduleR extends RealmObject {
         this.eventSetId = eventSetId;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
